chore: add winston types
