+++ package.json
@@ -31,6 +31,7 @@
     "@types/node": "^20.17.50",
     "@types/react": "^18.3.22",
     "@types/react-dom": "^18.3.7",
+    "@types/winston": "^2.4.4",
     "eslint": "^9.27.0",
     "eslint-config-next": "15.1.7",
     "husky": "^9.1.7",
